# perf: Refactor countScalar to benefit from simd